update kiểm tra ký tự

+++ app/src/main/java/com/example/learnforeignlanguage/LoginActivity.java
@@ -36,22 +36,26 @@ public class LoginActivity extends AppCompatActivity {
         editor = sharedPreferences.edit();
     }
     public void  login(View view){
-         list = userDao.timUserName(edt_userName.getText().toString());
-         if(list.size()==0){
-             Toast.makeText(this,"Tài khoản không tồn tại",Toast.LENGTH_LONG).show();
-         }else {
-                if(list.get(0).getPassword().equals(edt_pass.getText().toString())){
-                    int id = userDao.timUserName(edt_userName.getText().toString()).get(0).getIdUser();
-                    editor.clear();
-                    editor.putString("userName",edt_userName.getText().toString());
-                    editor.putInt("idUser", id);
-                    editor.commit();
-                    Intent intent = new Intent(this,MenuActivity.class);
-                    startActivity(intent);
-                }else {
-                    Toast.makeText(this,"Sai mật khẩu",Toast.LENGTH_LONG).show();
-                }
+         if(!(edt_userName.getText().toString().isEmpty()&&edt_pass.getText().toString().isEmpty())){
+             list = userDao.timUserName(edt_userName.getText().toString());
+             if(list.size()==0){
+                 Toast.makeText(this,"Tài khoản không tồn tại",Toast.LENGTH_LONG).show();
+             }else {
+                 if(list.get(0).getPassword().equals(edt_pass.getText().toString())){
+                     int id = userDao.timUserName(edt_userName.getText().toString()).get(0).getIdUser();
+                     editor.clear();
+                     editor.putString("userName",edt_userName.getText().toString());
+                     editor.putInt("idUser", id);
+                     editor.commit();
+                     Intent intent = new Intent(this,MenuActivity.class);
+                     startActivity(intent);
+                 }else {
+                     Toast.makeText(this,"Sai mật khẩu",Toast.LENGTH_LONG).show();
+                 }
 
+             }
+         }else {
+             Toast.makeText(this,"Không để trống thông tin đăng nhập",Toast.LENGTH_LONG).show();
          }
     }
     public void signUp(View view){
